+++ src/main/java/com/walkhub/walkhub/domain/auth/presentation/AuthController.java
@@ -1,0 +1,23 @@
+package com.walkhub.walkhub.domain.auth.presentation;
+
+import com.walkhub.walkhub.domain.auth.presentation.dto.request.SignInRequest;
+import com.walkhub.walkhub.domain.auth.presentation.dto.response.UserTokenResponse;
+import com.walkhub.walkhub.domain.auth.service.UserSignInService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
+
+@RequiredArgsConstructor
+@RestController("/users/auth")
+public class AuthController {
+
+    private final UserSignInService userSignInService;
+
+    @PostMapping
+    public UserTokenResponse userSignIn(@RequestBody @Valid SignInRequest request) {
+        return userSignInService.execute(request);
+    }
+}
