+++ src/main/java/com/walkhub/walkhub/domain/user/presentation/UserController.java
@@ -1,6 +1,6 @@
 package com.walkhub.walkhub.domain.user.presentation;
 
-import com.walkhub.walkhub.domain.auth.prsentation.dto.response.UserTokenResponse;
+import com.walkhub.walkhub.domain.auth.presentation.dto.response.UserTokenResponse;
 import com.walkhub.walkhub.domain.user.presentation.dto.request.UserAuthCodeRequest;
 import com.walkhub.walkhub.domain.user.presentation.dto.response.QueryMyPageResponse;
 import com.walkhub.walkhub.domain.user.service.QueryMyPageService;
